index/landing: Add a bit more of the functionality code
Now, with the nascent js/landing.js script, the page will load the image
in the background (into a cached buffer variable which is unused, called
"internalImage") before setting the "backgroundImage" property.

The reason of doing this is that browsers can reuse images when they
load them. So, if a browser loads the image in the background (as if it
were creating a new Image object) and then just changes the image url on
another object, that image will automagically get substituted in (after
a couple of continuity checks by the browser to ensure that it hasn't
changed)