Cache: Fix whitespace and add checking for non-truthy routes array
Now, Workers will *gracefully* whine if and when they cannot calculate
the array for routes. (For instance, if the effective filename is
something like #index.coffee#, which cannot be parsed as CoffeeScript as
it is an editor swapfile)

This works pretty well and will be an effective solution to this problem.

[GitHub] This commit resolves #47.